Flatten the tree history
